phase-4-advanced-architectures/day-21-production-rag-systems/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
requests==2.31.0
pydantic==2.7.1
python-dotenv==1.0.1
